Add book operations
