+++ main.rb
@@ -7,7 +7,7 @@ require './classes/book'
 require './classes/label'
 require './classes/author'
 require './classes/game'
-# require './operations/book_operations'
+require './operations/book_operations'
 # require './operations/music_operations'
 # require './operations/game_operations'
 # require './operations/file_operations'
